Make CI run the suite that would have caught the last four commits
The database job has been red since 0033 landed: the pgTAP file still
asserted the SPY benchmark block that migration deleted, and 0034's own
new file aborted on is(numeric, integer) — an integer literal where the
expression is numeric — losing five assertions to a bad plan. Four
commits shipped on top of that without anyone reading the job.

Both assertions are fixed, and 0032 now pins the ABSENCE of the block,
which is the property that has to hold: an edit re-adding the join would
otherwise pass every test and restore the payload that timed the
deployed reads out.

The unit suite and lint now run here as well. They never did, so the one
gate that could see a store regression was a job already failing for an
unrelated reason.

+++ supabase/tests/0034_market_home_blob.test.sql
@@ -145,7 +145,11 @@ select ok(
 
 select is(
   (select (public.market_home(array['ZTSTB'], '{}', '{}') -> 'rows' -> 0 ->> 'px')::numeric),
-  200,
+  -- 200::numeric, not 200. pgTAP resolves is() per argument type, and
+  -- is(numeric, integer, text) does not exist: an integer literal here aborts
+  -- the whole file with "function is(numeric, integer, unknown) does not
+  -- exist", which reads as a broken suite rather than a typing slip.
+  200::numeric,
   'and the new price is what the page gets'
 );
